# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Artichaut API",
        "description": "API documentation for Hotel Artichaut API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/user": {
            "get": {
                "tags": [
                    "AdminUsers"
                ],
                "summary": "Get all users- need to be authentified and role = employee",
                "operationId": "23b25d59267428953c5abfb5eb943cfe",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "post": {
                "tags": [
                    "AdminUsers"
                ],
                "summary": "add a new user- need to be authentified and role = manager",
                "operationId": "53b318761451e4ed7912fafc7a6994d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "firstname",
                                    "lastname",
                                    "address",
                                    "city",
                                    "postal_code",
                                    "phone",
                                    "role_id",
                                    "is_pro",
                                    "is_vip"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User's email address",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User's password (minimum 10 characters)",
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "description": "User's first name",
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "lastname": {
                                        "description": "User's last name",
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "address": {
                                        "description": "User's address",
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "city": {
                                        "description": "User's city",
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "postal_code": {
                                        "description": "User's postal code (5 digits)",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "User's phone number (10 digits)",
                                        "type": "string"
                                    },
                                    "role_id": {
                                        "description": "User's role",
                                        "type": "integer"
                                    },
                                    "is_pro": {
                                        "description": "User's status (optional)",
                                        "type": "boolean"
                                    },
                                    "is_vip": {
                                        "description": "Indicates if the user is VIP",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully created"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/admin/user/{id}": {
            "get": {
                "tags": [
                    "AdminUsers"
                ],
                "summary": "Get one user by id- need to be authentified and role = employee",
                "operationId": "da8a35cf9e338a35a7bc948a22a026bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "post": {
                "tags": [
                    "AdminUsers"
                ],
                "summary": "Update an existing user- need to be authentified and role = manager",
                "operationId": "3449a798d901929069ac7b8cb79e33a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "firstname",
                                    "lastname",
                                    "address",
                                    "city",
                                    "postal_code",
                                    "phone",
                                    "role_id",
                                    "is_pro",
                                    "is_vip"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User's email address",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User's password (minimum 10 characters)",
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "description": "User's first name",
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "lastname": {
                                        "description": "User's last name",
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "address": {
                                        "description": "User's address",
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "city": {
                                        "description": "User's city",
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "postal_code": {
                                        "description": "User's postal code (5 digits)",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "User's phone number (10 digits)",
                                        "type": "string"
                                    },
                                    "role_id": {
                                        "description": "User's role",
                                        "type": "integer"
                                    },
                                    "is_pro": {
                                        "description": "User's status (optional)",
                                        "type": "boolean"
                                    },
                                    "is_vip": {
                                        "description": "Indicates if the user is VIP",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully updated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "delete": {
                "tags": [
                    "AdminUsers"
                ],
                "summary": "Delete a user by id- need to be authentified and role = manager",
                "operationId": "5117178230ba4e8bc1cc2c813faf8679",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
...
    },
...
}